external_server/requirements.txt

annotated-types==0.7.0
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
dj-database-url==2.2.0
Django==5.0.6
djangorestframework==3.15.1
greenlet==3.0.3
idna==3.7
iniconfig==2.0.0
packaging==24.0
pluggy==1.5.0
psycopg==3.1.19
psycopg2-binary==2.9.9
pydantic==2.7.3
pydantic_core==2.18.4
pytest==8.2.1
requests==2.32.3
SQLAlchemy==2.0.30
sqlparse==0.5.0
typing_extensions==4.12.1
urllib3==2.2.1
coreapi==2.3.3
PyYAML==6.0.1

